imported tickets order there code from TicketOrderMobile.vue to TicketOrderThereTable.vue
